Move form styles rendering methods to style utils
[MAILPOET-2811]
