+++ lib/Form/Util/Styles.php
@@ -110,4 +110,64 @@ EOL;
     }
     return implode(PHP_EOL, $formattedStyles);
   }
+
+  public function renderFormDivWrapperStyles(array $form, string $selector = null): string {
+    if (is_null($selector)) return '';
+    if (!isset($form['settings'])) return '';
+    $formSettings = $form['settings'];
+    $styles = [];
+
+    if (isset($formSettings['backgroundColor'])) {
+      $styles[] = 'background-color: ' . trim($formSettings['backgroundColor']);
+    }
+
+    if (isset($formSettings['border_size']) && isset($formSettings['border_color'])) {
+      $styles[] = 'border: ' . $formSettings['border_size'] . 'px solid ' . $formSettings['border_color'];
+    }
+
+    if (isset($formSettings['border_radius'])) {
+      $styles[] = 'border-radius: ' . $formSettings['border_radius'] . 'px';
+    }
+
+    if (isset($formSettings['background_image_url'])) {
+      $styles[] = 'background-image: url(' . trim($formSettings['background_image_url']) . ')';
+      $backgroundPosition = 'center';
+      $backgroundRepeat = 'no-repeat';
+      $backgroundSize = 'cover';
+      if (isset($formSettings['background_image_display']) && $formSettings['background_image_display'] === 'fit') {
+        $backgroundPosition = 'center top';
+        $backgroundSize = 'contain';
+      }
+      if (isset($formSettings['background_image_display']) && $formSettings['background_image_display'] === 'tile') {
+        $backgroundRepeat = 'repeat';
+        $backgroundSize = 'auto';
+      }
+      $styles[] = 'background-position: ' . $backgroundPosition;
+      $styles[] = 'background-repeat: ' . $backgroundRepeat;
+      $styles[] = 'background-size: ' . $backgroundSize;
+    }
+    $media = "@media (max-width: 500px) {{$selector} {background-image: none;}}";
+
+    return $selector . '{' . join(';', $styles) . '}' . $media;
+  }
+
+  public function renderFormElementStyles(array $form): string {
+    if (!isset($form['settings'])) return '';
+    $formSettings = $form['settings'];
+    $styles = [];
+
+    if (isset($formSettings['fontColor'])) {
+      $styles[] = 'color: ' . trim($formSettings['fontColor']);
+    }
+
+    if (isset($formSettings['form_padding'])) {
+      $styles[] = 'padding: ' . $formSettings['form_padding'] . 'px';
+    }
+
+    if (isset($formSettings['alignment'])) {
+      $styles[] = 'text-align: ' . $formSettings['alignment'];
+    }
+
+    return join(';', $styles);
+  }
 }
