+++ tests/unit/Form/RendererTest.php
@@ -97,88 +97,4 @@ class RendererTest extends \MailPoetUnitTest {
     $recaptcha = $this->htmlParser->findByXpath($html, "//div[@class='mailpoet_recaptcha']");
     expect($recaptcha->length)->equals(0);
   }
-
-  public function testItShouldNotRenderStylesForFormWithoutSettings() {
-    $form = Fixtures::get('simple_form_body');
-    $styles = $this->renderer->renderFormElementStyles($form);
-    expect($styles)->equals('');
-  }
-
-  public function testItShouldRenderBackgroundColour() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['backgroundColor' => 'red'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('background-color: red');
-  }
-
-  public function testItShouldRenderFontColour() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['fontColor' => 'red'];
-    $styles = $this->renderer->renderFormElementStyles($form);
-    expect($styles)->equals('color: red');
-  }
-
-  public function testItShouldRenderBorder() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['border_size' => '22', 'border_color' => 'red'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('border: 22px solid red');
-  }
-
-  public function testItShouldRenderPadding() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['form_padding' => '22'];
-    $styles = $this->renderer->renderFormElementStyles($form);
-    expect($styles)->equals('padding: 22px');
-  }
-
-  public function testItShouldRenderAlignment() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['alignment' => 'right'];
-    $styles = $this->renderer->renderFormElementStyles($form);
-    expect($styles)->equals('text-align: right');
-  }
-
-  public function testItShouldRenderBorderWithRadius() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['border_size' => '22', 'border_color' => 'red', 'border_radius' => '11'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('border: 22px solid red;border-radius: 11px');
-  }
-
-  public function testItShouldRenderImageBackground() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['background_image_url' => 'xxx'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('background-image: url(xxx)');
-    expect($styles)->contains('background-position: center');
-    expect($styles)->contains('background-repeat: no-repeat');
-    expect($styles)->contains('background-size: cover');
-  }
-
-  public function testItShouldRenderImageBackgroundTile() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['background_image_url' => 'xxx', 'background_image_display' => 'tile'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('background-image: url(xxx)');
-    expect($styles)->contains('background-position: center');
-    expect($styles)->contains('background-repeat: repeat');
-    expect($styles)->contains('background-size: auto');
-  }
-
-  public function testItShouldRenderImageBackgroundFit() {
-    $form = Fixtures::get('simple_form_body');
-    $form['settings'] = ['background_image_url' => 'xxx', 'background_image_display' => 'fit'];
-    $form['styles'] = '.mailpoet_paragraph {}';
-    $styles = $this->renderer->renderStyles($form, 'prefix');
-    expect($styles)->contains('background-image: url(xxx)');
-    expect($styles)->contains('background-position: center top');
-    expect($styles)->contains('background-repeat: no-repeat');
-    expect($styles)->contains('background-size: contain');
-  }
 }
